split up files lib (#1685)
